requirements.txt
atomicwrites==1.4.1
attrs==22.2.0
black==23.1.0
certifi==2022.12.7
charset-normalizer==2.0.12
click==8.1.3
colorama==0.4.6
cssbeautifier==1.14.7
Django==2.2.16
django-behaviors==0.5.1
django-testdata==1.0.3
djhtml==3.0.2
djlint==1.19.16
EditorConfig==0.12.3
Faker==12.0.1
flake8==5.0.4
flake8-absolute-import==1.0.0.1
flake8-broken-line==0.6.0
flake8-commas==2.1.0
flake8-isort==6.0.0
flake8-pie==0.16.0
flake8-plugin-utils==1.3.2
flake8-quotes==3.3.2
flake8-return==1.2.0
html-tag-names==0.1.2
html-void-elements==0.1.0
idna==3.4
iniconfig==2.0.0
isort==5.12.0
jsbeautifier==1.14.7
mccabe==0.7.0
mixer==7.1.2
mypy-extensions==1.0.0
packaging==23.0
pathspec==0.11.0
pep8-naming==0.13.3
Pillow==8.3.1
platformdirs==3.0.0
pluggy==0.13.1
py==1.11.0
pycodestyle==2.9.1
pyflakes==2.5.0
pytest==6.2.4
pytest-django==4.4.0
pytest-pythonpath==0.7.3
python-dateutil==2.8.2
pytz==2022.7.1
PyYAML==6.0
regex==2022.10.31
requests==2.26.0
six==1.16.0
sorl-thumbnail==12.7.0
sqlparse==0.4.3
toml==0.10.2
tomli==2.0.1
tqdm==4.64.1
typing_extensions==4.5.0
urllib3==1.26.14
